Process maps in worker
The interactive cancelation is not good enough to keep the service
responsive for queries, so use worker pool with queue.
